Added CB Model Research
Work on a content-based model for student-tutor matching was suggested. The files in the folder include an initial MatLab version, a revised python edition, and a helper file to load user.json data. If you run loadTutors, it prints out cbModelPython's output. This can give a good example of what the cbModel's power can be used for!